remove gisaid acknowledgements from ns tree create modal
