+++ src/frontend/src/views/Data/components/SamplesView/components/SampleTableModalManager/components/CreateNSTreeModal/components/Acknowledgement/acknowledgementConfig.ts
@@ -1,9 +1,9 @@
 import { Pathogen } from "src/common/redux/types";
 import { PathogenConfigType } from "src/common/types/pathogenConfig";
-import { GisaidAcknowledgement } from "./components/GisaidAcknowledement";
+import { NSAcknowledgement } from "./components/NSAcknowledement";
 import { NcbiVirusAcknowledgement } from "./components/NcbiVirusAcknowledgement";
 
 export const AcknowledgementConfig: PathogenConfigType<() => JSX.Element> = {
-  [Pathogen.COVID]: GisaidAcknowledgement,
+  [Pathogen.COVID]: NSAcknowledgement,
   [Pathogen.MONKEY_POX]: NcbiVirusAcknowledgement,
 };
