+++ src/frontend/src/views/Data/components/SamplesView/components/SampleTableModalManager/components/CreateNSTreeModal/components/Acknowledgement/components/NSAcknowledement/index.tsx
@@ -1,0 +1,9 @@
+import { Attribution, NextstrainLogo } from "./style";
+
+export const NSAcknowledgement = () => {
+  return (
+    <Attribution>
+      Built in partnership with <NextstrainLogo />.
+    </Attribution>
+  );
+};
