--- src/frontend/src/views/Data/components/SamplesView/components/SampleTableModalManager/components/CreateNSTreeModal/components/AcknowledgementFooter/index.tsx
@@ -1,9 +1,0 @@
-import { useSelector } from "react-redux";
-import { selectCurrentPathogen } from "src/common/redux/selectors";
-import { AcknowledgementFooterConfig } from "./acknowledgementFooterConfig";
-
-export const AcknowledgementFooter = (): JSX.Element | null => {
-  const pathogen = useSelector(selectCurrentPathogen);
-  const Component = AcknowledgementFooterConfig[pathogen];
-  return Component ? <Component /> : null;
-};
